support mongo memory server v7 M1 processor api changes, refactor over existing tests
