+++ monolith/package.json
@@ -43,7 +43,7 @@
         "@types/request": "^2.48.5",
         "@types/request-promise-native": "^1.0.17",
         "jest": "^26.6.3",
-        "mongodb-memory-server": "^6.9.3",
+        "mongodb-memory-server": "^7.4.0",
         "ts-jest": "^27.0.3",
         "tslint": "^5.20.1",
         "typescript": "^3.9.10"
